refactor: HOC 패턴으로 리팩토링
