+++ src/components/timetableComponents/hooks/useSlotSelection.ts
