+++ src/components/timetableComponents/parts/ColumnTitle.tsx
@@ -1,8 +1,7 @@
+import { DateType } from '../Timetable';
 import Text from 'components/atomComponents/Text';
 import styled from 'styled-components';
 import { theme } from 'styles/theme';
-
-import { DateType } from '../Timetable';
 
 interface ColumnTitleProps {
   availableDates: DateType[];
