+++ src/components/timetableComponents/parts/Slot.tsx
