+++ src/components/timetableComponents/parts/SlotTitle.tsx
