+++ src/components/timetableComponents/utils.ts
